feat(sources): lsp: check nil before indexing info.data.client_id (#13)
* Feat: make LspAttached callback more robust

There seem be cases where a `LspAttach` event does not contain a `data`
property with the client identifier inside. In such case the callback
function would fail as it simply assumes this data structure to be
given.
As it's still unclear why such events happen, it is quite easy to make
the callback more robust. Therefore it checks if the data exists and it
also verifies if a client could be found. In case that not, it simply
does nothing and does not attach to it.

* feat(sources): lsp: check nil before indexing info.data.client_id

* style: format with stylua

---------

Co-authored-by: Zeng <76579810+Bekaboo@users.noreply.github.com>