fix(menu): highlighting issues that occur during menu navigation (#52)
* fix(menu): preview doesn't update on menu reopen

- remove the check if cursor is same as prev_cursor field, which
  prevents preview updating when returning to an existing menu window

* fix(menu): context highlight not restored when submenu closed

When a submenu is closed and the parent menu regains focus, restore the
context highlight if it is set (i.e. if the cursor field is set).
Otherwise clear the context highlight.

* fix(menu): hover highlight broken if quick navigation disabled

In CursorMoved autocmd, the prev_cursor field is only updated through
the quick_navigation function. The hover highlight update takes the
prev_cursor field as an argument and so prior to this fix, the hover
highlight would only be updated when the quick navigation config option
was enabled